refactor(tests): moved '__tests__' folder inside 'src' and restructured it
