+++ src/__tests__/components/centeredWrapper.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import {render} from '@testing-library/react';
-import CenteredWrapper from '../src/components/layouts/CenteredWrapper';
+import CenteredWrapper from '../../components/layouts/CenteredWrapper';
 
 describe('Testing CenteredWrapper',() => {
     test('renders without errors', () => {
