+++ src/__tests__/pages/home.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import Home from '../src/pages/index';
+import Home from '../../pages/index';
 
 describe('Home', () => {
     test('renders a heading', () =>{
